Eliminate `resizing` in C API `Execute`.
The resizing of `c_output_lists{, _storage}` causes significant overhead when enabling the C API.

In this change, we initialize these variables to have the correct size first-time around, saving us the repeated `resize`ing. This comes at the expense of a small increase in code duplication in `Execute` and `ExecuteWithSingleDevice`, since the code is no longer shared by `GetCommonExecuteArgs`.

PiperOrigin-RevId: 755575548