[XLA:GPU] Do not remove duplicate arguments from kernel interface.
Currently XLA emitters pipeline will change entry function signature if some operands are duplicated. To keep track of the change, we need to recreate and use a map of argument indices before and after the pipeline in multiple places in the codebase. It's much easier to relink duplicate argument to one function arg and keep other args unused.

PiperOrigin-RevId: 788405826