Unify how AutotuneCacheKeys are getting constructed
Instances of `AutotuneCacheKey` could be constructed from a `model` and a `instruction` string. This makes it very easy for users to inject arbitrary strings into the cache key and makes it impossible for the owners of this class to maintain
a unified cache key structure.

This change reduces the 4 constructors down to 1 (public) constructor which takes an `HloInstruction` and a `DeviceDescription`. From the two object the constructor generates the cache key string itself keeping control over how the cache key is structured.

In addition some minor clang-tidy fixes in the touched files.

This change is not changing the format of the cache key, but I have no way of proving that, therefore I'm bumping the cache version to be and the safe side.

PiperOrigin-RevId: 790692701