Simplify coordination service identification of restarted out_of_sync workers to exclude in barrier.
Always mark a worker out of sync after restart and use it to bypass barrier check for out of sync workers.

Previously we only marked out of sync if the worker was waiting on some other barrier. This simplifies the logic and also excludes the out of sync worker earlier.

PiperOrigin-RevId: 791746924