Modify BlockLevelEmitter to be run before any triton config is found & use the CostModel to generate a default config
This will allow us to autotune between the "best" BlockLevelEmitter config and other backends without destroying compilation time. This will happen before any triton config is ever created so it should check if it's feasible to go through triton at all & then attempt to create a config for it. If an error arises or no config is found, this should be handled gracefully allowing the autotuner to simply not use it as a backend.

PiperOrigin-RevId: 802126971