[XLA:GPU] Enable autotuner for the scaled dot fusions.
We extend the autotuner to the scaled-dot case.
The autotuner has a very restrictive set of config for the scaled dot due to the issues of the scaled-dot -> regular dot AccelerateMatmul rewriter of Triton.

The numerics is correct for the config that actually works but requires a transpose for the rhs scales. It seems it is a bug of some other Triton rewrite.

PiperOrigin-RevId: 809928192