Fix HloValueAnalysis to work with the latest attribute  _xla_host_transfer_rendezvous.
Previously, the pass expects a Send/Recv pair to have the same string value in attribute _xla_host_transfer_rendezvous, this is not the same case anymore. We now have to remove the substring args_dtoh_ or retvals_htod_ from the attribute value to find the matching Send/Recv pair.

PiperOrigin-RevId: 810097899