[XLA:GPU] Move code out of coalescing_analysis.cc (NFC)
We want to reuse this code for determining possible unroll factors. Since
coalescing_analysis depends on fusion emitters, and we want to compute unroll
factors in fusion emitters, we need to make sure the extracted code has no
dependency on fusion emitters. Therefore we need to also adapt the function
signature of GetThreadIdToInputMemoryLayoutsMaps.

PiperOrigin-RevId: 810785811