[IFRT] Add `WithUserContext` and `WithCurrentUserContext` wrapper
`WithUserContext` takes a function and a `UserContextRef` and sets it up with `UserContextScope` when the returned function is called.

`WithCurrentUserContext` is a shorthand for `WithUserContext` that captures `UserContextScope::current()`.

This is a helper for handing over the user context into a function or callback that runs on a different thread or in a deferred way.

PiperOrigin-RevId: 813015441