Add `pjrt/c_api_client` directory.
This new directory contains the thick client wrapper that provides a C++ (e.g., `PjRtClient`, `PjRtDevice`) on top of the purely C-style PJRT C API. The `README.md` details the classes implemented to wrap the C API structures. This existed previously, generating this directory to improve organization and prepare for splitting these dependencies out from a monofile in further cls.

PiperOrigin-RevId: 813805944