Move GpuNormConfig creation from the IR Emitter to the `NormThunk` construction
This is refractor in preparation for the [de]serialization of the `NormThunk`. We want to [de]serialize based on the `GpuNormDescriptor` instead of the `GpuNormConfig`, but for that we need to keep a reference of the `GpuNormDescriptor` in the Thunk.

Considered creating the `GpuNormConfig` on the fly when needed, but in that case we'd need to create it during the `ExecuteOnStream` method, which IIUC we want to avoid for performance reasons.

PiperOrigin-RevId: 816110696