Introduce dumping of thunk metadata to file
This change makes XLA dump metadata information about the thunk graph into a proto file when the user requests an XLA dump.

The metadata is meant to be used with the checksum tracing feature. Long term the plan is to replace the Metadata dump by the more comprehensive thunk-to-proto serialization-feature but since this is not yet supported for all thunk kinds, just dumping metadata - which is thunk kind agnostic - is the stopgap at hand.

PiperOrigin-RevId: 816580387