[XLA:GPU] Add Thunk::TransformAllNestedThunks
A function similar to ForAllThunksMutable, but capable of wrapping or replacing nested thunks with new ones. In case the thunk has some specific requirements (e.g. assumes the directly nested thunk is a `SequentialThunk`), the implementation needs to ensure the assumptions still hold. The transform function must be infallible, but it may be a no-op and return the original `unique_ptr<Thunk>`.

The use for this is in buffer checksumming: to cover all thunks, we need to recursively insert pre-/post-execution checksum thunks around execution of nested thunks. Unfortunately, ForAllThunksMutable doesn't allow mutating the thunk's *container*, which would create the need to switch() over thunk kind, special-case any thunk that may contain nested thunks, and expose the internals to the caller to reimplement the recursion.

PiperOrigin-RevId: 825550682