[XLA:GPU]: Add ability to pass opaque args count to xtile entry function.
In certain cases the arguments to a triton/xtile function could be scalars
or opaque arguments such as pointers to synchronization memory on a another GPU
that should not be tiled with the regular tiling infra.

This change allows skipping type validation for such arguments if the count
of such arguments is explicitly specified as an attribute.

PiperOrigin-RevId: 827433879