Add `xla_cpu_enable_platform_dependent_math` flag
This mirrors the backend flag `disable_platform_dependent_math`. This flag is currently always set, and I don't think we should change that in this change, as it would be a breaking change for clients that haven't had a chance to set this new front end flag yet.

This is currently hooked up to YNNPACK in this change, however, it is not currently tested, because:
- YNNPACK fp32 and bf16 dots are currently disabled (in part due to lacking this flag).
- XLA dots currently are already not deterministic (unless using backend specific environment variables).
- Therefore, this flag has no functional changes that can be observed with any available options.

I think that we should add the flag, allow clients to set the flag, then we can enable YNNPACK, and then we can add some test coverage of this behavior. We could add some more flags to control whether fp32 and bf16 is enabled in YNNPACK, but I don't think we should do that.

For now, I've manually tested this with fp32 and bf16 support in YNNPACK hacked on, and verified that this flag does cause YNNPACK to avoid using `tile_k > 1` kernels (the only behavior currently affected by this flag).

PiperOrigin-RevId: 827801469