[SymbolicMap] Add basic simplifications for Add and Mul when we create them
Added basic simplifications to + and * operators to mimetize what we currently have in AffineMap. Sadly IndexingMapTests are a bit contradictory and rely too much in that simplifications. We will use our own canonicalization after fully merging.

To be clean, I introduced SimplifyMulByConstantRHS: A new helper function to encapsulate the simplification rules for multiplications where the right-hand side is a constant. This includes constant folding, handling neutral elements (0 and 1) and associativity.

Adapted some tests for SymbolicExpr and SymbolicMap so these simplifications are not triggered.

PiperOrigin-RevId: 828483457