Use KernelArgumentsPackingSpec in KernelSpec
This integrate `KernelArgumentsPackingSpec` into `KernelSpec`. `KernelSpec` can now either carry a `KernelArgsPackingFunc` (which is the previously existing callback) or the new `KernelArgumentsPackingSpec`. The latter will turned into a `KernelArgsPackingFunc` during kernel loading (while creating a `Kernel` object).

`KernelSpec` instances with a custom `KernelArgumentsPackingSpec` can now be serialized using `KernelSpec::ToProto`.

The feature gets a unit test for `KernelSpec` and an integration in `gpu_kernel_test.cc`. The latter verifies that the whole flow (loading and executing a kernel with a custom `KernelArgumentsPackingSpec`) works.

PiperOrigin-RevId: 829388030