Support serialization of InprocessSymbolSpecs
`KernelLoaderSpec::InprocessSymbol` reference CUDA or ROCm kernels through a function pointer. This function pointer is only valid in the currently running process and can't easily be serialized for usage in a different process.

Therefore this change introduces the `KernelSymbolRegistry` which is a mapping from a unique name (string) to a function pointer. With the help of this registry we can use the name for serialization and recover the function pointer in a different process.

PiperOrigin-RevId: 831274895