[XLA:GPU] Remove the memcpy-based implementation of RaggedAllToAll.
Memcpy-based implementation didn't show any benefits over other implementations. One-shot kernel + multi-host decomposer should be used in most of the production setups that have NVLink. NCCL can be used as slow, but the most generic implementation, even for multi-host. Memcpy-based returned out to be the slowest out of all and it only works on single host.

PiperOrigin-RevId: 837128507