Move SymbolicExprStorage registration out of expression creation.
This CL moves the registration of SymbolicExprStorage out of GetOrCreateSymbolicExpr and into the creation site of the MLIRContext. This avoids repeated registration for every expression and improves performance. Ideally we should register SymbolicExprStorage just once, but as we are creating and using MLIRContext in many many different places, we need to Register in each of this places to avoid crashing.

Compile time benchmarks (rebasing on top of cl/802100018)
Previous: 0.93x (http://gpaste/5608727887609856)
After: 0.98x (http://gpaste/4716749884882944)
PiperOrigin-RevId: 838665853