[XLA] Avoid duplicated include paths in XLA build
The change that introduces googletest wrapper does not work reliably:
depending on the order of deps in BUILD file, googletest include paths
may end up before the wrapper ones. When this happens, the
ASSERT_OK/EXPECT_OK macros are not available.

This showed up in a bzlmod build once I started migrating the TF_
macros.

I tried multiple ways to fix it, described in
README.add-status-macros.md for anyone trying to make this nice in the
future. In the end, adding the macros directly to googletest in a patch
seems like the least intrusive option.

PiperOrigin-RevId: 839773153