Remove cyclic dependency between AotCompilationResult and Compiler
`AotCompilatioResult::LoadExecutable` used to take a pointer to `Compiler` which leads to a cyclic dependency since `Compiler::CompileAheadOfTime` returns instances of `AotCompilationResult`.

The pointer to compiler is only needed by the legacy GPU AOT compilation flow and can be worked around by storing the pointer in the object when `LegacyGpuAotCompilationResult` gets created.

PiperOrigin-RevId: 841647395