[XLA:GPU] Run DotMerger after the simplification fix-point pass pipeline.
The simplification passes alg-simplifier, constant_folding, cse, and dot-strength-reduction are cheap compared to dot-merger. Running these passes to a fix-point first can greatly reduce the number of merge candidates that dot-merger needs to consider. The dot-merger pass itself is designed to do all merging in one go. Hence, there is no need to run it to a fix point.

Also clarify the semantic of the dot-merger threshold in xla.proto and remove an early exit in dot_merger.cc that was not in-line with it.

PiperOrigin-RevId: 845245018