Remove xtile_compiler stub and simplify BUILD rules.
This change removes the conditional compilation for `xtile_compiler` and the associated stub implementation. The real implementation is now always used, simplifying the BUILD file structure.

I'm not sure why the stub was introduced in the first place. All usages of xtile_compiler are in GPU-only code anyway. I suspect that it might have been that the `compilation_pipeline` target used to depend on CUDA headers or something like that and not having the stub was breaking the deviceless unit tests.

PiperOrigin-RevId: 848445454